Speed up our library loader.
I noticed that the previous versions of the library were calling the find_module() method a redundant number of times, so we're using the new Modules::find() method to get the exact path and unshift that to the top of the places to look for the library.

We could boost performance even a touch more if we simply loaded the file ourself here, like WireDesignz did, but we'll leave it like this for the time being.