Translate Module Slow w/ large # of modules
- Assets::add_module_js() should only be called from the controller's
constructor
- Modules::files() was extremely slow when searching for language files
in all modules (especially when there are a large number of application
modules). This was caused by the function mapping the whole modules
directory (either application or bonfire) if the module was not found in
the current path. This should only be done if the $module_name is not
passed.