Improved docs configuration, fix sidebar for app/dev docs when not using a toc file
- The sidebar was generating links starting with
'application/application/' or 'developer/developer/' if a toc file
wasn't used to specify the links. The duplication was caused by a
modification made to allow module documents to reside in either the
application or bonfire directory. This has been resolved. See changelog
for known issue resulting from this change.
- Added 'docs.toc_file' config option to allow changing the name of the
toc file (though it still has to use the same format). It will default
to '_toc.ini' if the option is not present.
- Added 'docs.permitted_environments' config option, which is an array
of environments in which the docs will be displayed. If the config entry
is not present or the array is empty, the docs module will display as
usual. If one or more environments are listed in the array and the
current environment is not listed, the docs module will redirect the
user to the site's base URL.