Fix #757 and other Migrations updates
- Re-arranged changelog slightly, see the Migrations section under
"Additional Changes" for a more detailed listing of the changes to the
Migrations module.
- Fix #757: version is always up when SQL migrations fail
- Fixed an error in the library's version() method when an error
occurred in glob() while retrieving the migrations for a given step
- Made numerous changes to reduce the number of database hits,
especially when auto-migrating, when a large number of modules are
installed, or when migrating multiple versions at once.
- Overhauled the library's API, deprecated most of the existing public
methods/properties (all public methods/properties still work, but may
not perform quite as well, since the methods are forwarding requests to
the new methods) - I intend to update any other uses of the library in
Bonfire's code to make sure the deprecated methods are not used
elsewhere.