timezone_menu accessibility
- Added timezone_menu function to BF_date_helper to override the CI
date_helper's version of the function.
- Added 4th parameter to timezone_menu() to allow passing additional
attributes for the generated select element.
- This function should be removed from BF_date_helper if/when CI is
updated to a version that includes this change in its date_helper (it
has been in the develop branch for almost 2 years).
- See @internal note in function header for additional information.
- Updated user_fields view to use the updated function to add an ID
matching the for attribute on the label (which fixes an accessibility
error for all of the forms that use these fields, since the for
attribute must reference an ID value)
- Related to #505 and #538 (both of which were closed at the time of
this commit)