Update Migrations.php
Fixing logical error in conditional on line 945 where trying to check if var $currentVersion is "empty()" when that version is set to zero (0) to check if $currentVersion is !== '0' instead. PHP empty() function would consider zero to be false, and thus empty and then the subsequent SQL is applied incorrectly as a result. This patch resolves that issue and allows the installer to proceed properly now.