Fix #1048 (almost) complete Russian language files
Russian language files for Bonfire Modules and updates for application
language files (and translate module) as submitted by @cjmaxik with some
changes to formatting and comment headers.

install_lang.php in the submitted zip file was empty, so the only
changes made to that file were to update the header and formatting.