Revert to using MX_Loader and MX_Controller, but keep updated Router and Modules functionality
- Set Base_Controller to extend MX_Controller, remove other changes
related to the loader changes.
- Revert MX/Modules and MX/Router to current versions from Wiredesignz
repo (these files are no longer used)
- Revert changes in Codeigniter
- Revert changes in BF_Lang
- Add BF_Loader (extends MX_Loader) to port some of the changes (Sparks,
loading helpers with the BF_ prefix from the /bonfire/helpers/
directory, and loading libraries with the BF_ prefix from the
/bonfire/libraries directory)
- Added BF_Router to take the place of MX_Router, load
/bonfire/libraries/Modules.php instead of
/application/third_party/MX/Modules.php, and load
/bonfire/libraries/Route.php to permit use of new routing features. Also
should fix #1070.
- Removed /bonfire/core/Loader.php and /bonfire/core/Router.php
- Made Modules library closer to the MX Modules class to reduce the
likelihood of errors like #1070.
- Reverted changes to view loading in the template library which don't
work with the CI/MX Loader. TODO: add functionality to the BF_Loader to
allow this to work without using methods and properties on the CI Loader
which will likely be private/protected in the future.
- Some /public/index.php changes ported from CI3.