Fix #337 Can not log in to admin when site is turned off
Issue was previously closed as documentation/non-issue.

Added a variable (`$allowOffline`) to App_hooks which allows the pages
in the array assigned to the variable to bypass the site offline mode.
The format of the list is the same as the `$ignore_pages` variable in
the same file.

If the array contains `'/users/login'`, users with the correct
permissions will be able to login to the site as usual. If they do not
have the correct permissions, once they login the site will display the
offline error page. If you need to disable user login while the site is
offline, set `$allowOffline` to an empty array or remove the
`'/users/login'` and `'/users/logout'` entries from the array (if you
have added other pages which you want to permit to remain online).