Fix #1097, #1081 Installer looks for development database config
Updated installer lib to look for config/ENVIRONMENT/database.php, then
fallback to config/development/database.php and config/database.php.

Note that it will simply load the first one it finds when searching in
this order, so it will still fail if the first of the above is invalid
for the current environment.