Initial testing of CI3, leaving CI2 support intact
To use CI3, copy the CI3 system files into bonfire/ci3 (do not overwrite
the files included in this folder by Bonfire), then open
public/index.php and replace the $system_path setting with the CI 3
system path setting (for ease of use, the CI3 $system_path is included
in the file, commented out).

This version of Bonfire is NOT ready for production use when CI3 is
enabled (but CI2 should still work fine).