Add ci3_sessions table for database sessions driver support in CI3
- Added migration to create ci3_sessions table, so CI3 with database
sessions and CI2 can be used from the same installation (primarily for
testing changes required for CI3 at this time)
- Added CI_VERSION checks to calls to load the session library/driver
- Added doctype to index.html files
- Updated config to set the sessions driver to database and set the
sessions table to ci3_sessions (CI2 uses a different config value to set
its table name, so this does not change the table used by CI2)

Note that CI3 will not work in this configuration until the migration
has been run or the database table has been added manually (the
migration allows for this by not creating the table if it already
exists). So, you can either run the migration under CI2 (which should
not cause any problems) or configure a different session driver.