Updated CommonMark Library and Drivers
Apologies to anyone that already implemented a driver for this. Let me
know if you need any help updating your driver to work.

If you're using the CommonMark library for something other than the docs
module, you'll need to update your code to call
$this->commonmark->convert() instead of $this->commonmark->parse(),
though the latter will continue to work for a while.